requirements.txt
xmltodict==0.10.2
PyYAML==3.11